Support removed operations in `definitionToEndpoint`
This repository uses data generated and stored in `octokit/openapi`.
Where one API version has an operation and another doesn't, then
we add a special operation entry representing the missing operation.
This is done in the [`addRemovedOperations` method][1].

When these special operation entries are added, they doesn't have
the `X-GitHub` extension data included. Currently, we expect that
to be there and blow up if it isn't. This makes the code more tolerant.

[1]:
https://github.com/octokit/openapi/blob/65f7b1ce4f0d90fa1db433d7f0904ecd6cc3bdc4/scripts/build.js#L413