feat: use the `ghec` schema bundle for querying for `octokit/openapi` versions v8.0.0 onwards
As of v8.0.0 of `octokit/openapi`, the `api.github.com` bundle
does not include API operations exclusive to GitHub Enterprise
Cloud. If you want to see the GHEC-exclusive operations and all
of the other normal ones, you have to look at the `ghec` bundle.

This points our GraphQL queries to use the `ghec` bundle if the
requested version is on or after v8.0.0. This will ensure that
all expected operations are returned, and you can still filter
them to find the ones you care about using the `githubCloudOnly`
input.