feat: maintain compatibility with `octokit/openapi` `v8.0.0` and later by querying against the `ghec` schema bundle (#78)
* Support removed operations in `definitionToEndpoint`

This repository uses data generated and stored in `octokit/openapi`.
Where one API version has an operation and another doesn't, then
we add a special operation entry representing the missing operation.
This is done in the [`addRemovedOperations` method][1].

When these special operation entries are added, they doesn't have
the `X-GitHub` extension data included. Currently, we expect that
to be there and blow up if it isn't. This makes the code more tolerant.

[1]:
https://github.com/octokit/openapi/blob/65f7b1ce4f0d90fa1db433d7f0904ecd6cc3bdc4/scripts/build.js#L413

* feat: use the `ghec` schema bundle for querying for `octokit/openapi`
versions v8.0.0 onwards

As of v8.0.0 of `octokit/openapi`, the `api.github.com` bundle
does not include API operations exclusive to GitHub Enterprise
Cloud. If you want to see the GHEC-exclusive operations and all
of the other normal ones, you have to look at the `ghec` bundle.

This points our GraphQL queries to use the `ghec` bundle if the
requested version is on or after v8.0.0. This will ensure that
all expected operations are returned, and you can still filter
them to find the ones you care about using the `githubCloudOnly`
input.

Co-authored-by: Gregor Martynus <39992+gr2m@users.noreply.github.com>